refactor: split generation service file
