+++ server/src/core/engine/generation/index.ts
@@ -1,0 +1,3 @@
+﻿export { generateProjectFile } from './build'
+export { applyProjectEdit, applyProjectFix } from './changes'
+export { extractProjectIntent, extractProjectPlan } from './intake'
